requirements.txt
Django>=5.2.10
djangorestframework>=3.16.1
pandas>=2.3.3
openpyxl>=3.1.5
requests>=2.32.5
